x,y,z to doubles and dont draw on menu
made x, y, and z to doubles from floats so that I could have values in my map that are size max int just in case. Also made it so the program doesnt write pixels over the menu. Just nicer visually.

+++ fdf_working/includes/fdf.h
@@ -6,7 +6,7 @@
 /*   By: dpentlan <dpentlan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/29 10:36:13 by dpentlan          #+#    #+#             */
-/*   Updated: 2023/06/28 07:58:59 by dpentlan         ###   ########.fr       */
+/*   Updated: 2023/06/28 09:48:48 by dpentlan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -73,9 +73,9 @@
 //	Sub Structs
 typedef struct s_point
 {
-	float			x;
-	float			y;
-	float			z;
+	double			x;
+	double			y;
+	double			z;
 	int				x_coord;
 	int				y_coord;
 	int				color;
